Fixed importing errors and general console bugs
